meeting-voice-assistant/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiohttp>=3.9.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydub>=0.25.0

# LangChain / LangGraph for audio analysis
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
